chore(Standardization): creating default export file
